change : double to decimal for better precision

+++ backend/api/Dtos/Responses/AmortizationScheduleSummaryResp.cs
@@ -4,8 +4,8 @@ namespace LoanApi.Dtos.Outputs
 {
     public class AmortizationScheduleSummaryResp
     {
-        public double MonthlyPaymentAmount { get; set; }
-        public double TotalInterestPaid { get; set; }
+        public decimal MonthlyPaymentAmount { get; set; }
+        public decimal TotalInterestPaid { get; set; }
         public IList<Payment> Payments { get; set; }
     }
 }
